Add list generating and differencing tool
Rename forecast_stats.py to forecast_list.py.  This tool now
computes statistics and lists files that match user command-line
filters.  Users can specify a range of months, a single month,
a list of models, and a model flavor.  An --exclude list makes
the tool only print/count files that do NOT appear on a given
list.

This tool can be used to generate lists of new forecasts to be
processed by SPHINX.