Define models using model_root keys
Previous commits forgot to add the new models to the `models`
list.  Now defining `models` using the keys of `model_root`,
to prevent future mismatches.  The consequence of this is that
`models` will be forced to alphabetical order and will not
omit any models.  Cannot think of any downside to that at the
moment.